requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0